Don't assumme all APN changes require cleanup
This patch checks that we need to actually cleanup connections
when calling tryRestartDataConnections.  This is necessary because
when a user goes to add a new APN it tears down the data connection
and then restarts it.  The same is true when the user discards their
changes.  The only time we should be restarting data connections is
when the user has actually switched APNs.

Change-Id: I48fa3044d3d79ee0bdde16557bd65eed6c4b7254
TICKET: CYNGNOS-988